JAMES-3492, move classes to separate package
