+++ server/container/metrics/metrics-es-reporter-v7/src/main/java/org/apache/james/metrics/es/v7/ESMetricReporter.java
@@ -17,7 +17,7 @@
  * under the License.                                           *
  ****************************************************************/
 
-package org.apache.james.metrics.es;
+package org.apache.james.metrics.es.v7;
 
 import java.io.IOException;
 import java.util.Optional;
