+++ server/container/metrics/metrics-es-reporter-v7/src/test/java/org/apache/james/metric/es/v7/ES6ReporterTest.java
@@ -17,15 +17,13 @@
  * under the License.                                           *
  ****************************************************************/
 
-package org.apache.james.metric.es;
+package org.apache.james.metric.es.v7;
 
 import org.apache.james.backends.es.v7.DockerElasticSearch;
-import org.apache.james.util.docker.Images;
 import org.junit.jupiter.api.extension.RegisterExtension;
 
-class ES2ReporterTest extends ESReporterContract {
+class ES6ReporterTest extends ESReporterContract {
 
     @RegisterExtension
-    static DockerElasticSearchExtension testExtension = new DockerElasticSearchExtension(
-        new DockerElasticSearch.NoAuth(Images.ELASTICSEARCH_2));
+    static DockerElasticSearchExtension testExtension = new DockerElasticSearchExtension(new DockerElasticSearch.NoAuth());
 }
