+++ server/container/metrics/metrics-es-reporter-v7/src/test/java/org/apache/james/metric/es/v7/ESReporterContract.java
@@ -17,7 +17,7 @@
  * under the License.                                           *
  ****************************************************************/
 
-package org.apache.james.metric.es;
+package org.apache.james.metric.es.v7;
 
 import static io.restassured.RestAssured.when;
 import static org.awaitility.Awaitility.await;
@@ -31,8 +31,8 @@ import org.apache.james.backends.es.v7.DockerElasticSearch;
 import org.apache.james.metrics.api.Metric;
 import org.apache.james.metrics.api.TimeMetric;
 import org.apache.james.metrics.dropwizard.DropWizardMetricFactory;
-import org.apache.james.metrics.es.ESMetricReporter;
-import org.apache.james.metrics.es.ESReporterConfiguration;
+import org.apache.james.metrics.es.v7.ESMetricReporter;
+import org.apache.james.metrics.es.v7.ESReporterConfiguration;
 import org.awaitility.Duration;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
